+++ rollup.config.api.js
@@ -5,6 +5,7 @@ const terser = require('@rollup/plugin-terser').default;
 const json = require('@rollup/plugin-json');
 const nodePolyfills = require('rollup-plugin-polyfill-node');
 const externalGlobals  = require('rollup-plugin-external-globals');
+const dotenv = require('rollup-plugin-dotenv').default;
 
 const options = {
     input: 'out/api/view.js',
@@ -12,8 +13,8 @@ const options = {
         file: 'api.view.js',
         format: 'iife'
     },
-    external: ['react', 'react-dom', 'prismjs', 'react-markdown'],
     plugins: [
+        dotenv(),
         externalGlobals({
             'react': 'React',
             'react-dom': 'ReactDOM',
@@ -39,6 +40,12 @@ const options = {
             ]
         }),
         terser()
+    ],
+    external: [
+        'react',
+        'react-dom',
+        'prismjs',
+        'react-markdown'
     ]
 };