+++ rollup.config.node.js
@@ -3,7 +3,7 @@ const { nodeResolve } = require('@rollup/plugin-node-resolve');
 const terser = require('@rollup/plugin-terser').default;
 const json = require('@rollup/plugin-json');
 const nodePolyfills = require('rollup-plugin-polyfill-node');
-const externalGlobals  = require('rollup-plugin-external-globals');
+const dotenv = require('rollup-plugin-dotenv').default;
 
 const options = {
     input: 'out/extension.js',
@@ -14,17 +14,12 @@ const options = {
         format: 'cjs',
         globals: {
             'vscode': 'vscode',
-            'path': 'path'
+            'path': 'path',
+            'https': 'https'
         }
     },
     plugins: [
-        externalGlobals({
-            'react': '{"$":"react"}',
-            'react-dom': '{"$":"react-dom"}',
-            'prismjs': '{"$":"prismjs","languages":{}}',
-            'react-markdown': '{"$":"react-markdown"}',
-            'react-in-viewport': '{"$":"react-in-viewport"}',
-        }),
+        dotenv(),
         json(),
         commonjs({
             esmExternals: ['vscode', 'path'],
@@ -36,7 +31,11 @@ const options = {
         }),
         terser()
     ],
-    external: ['vscode', 'path', 'react', 'react-dom', 'prismjs', 'react-markdown', 'react-in-viewport']
+    external: [
+        'vscode',
+        'path',
+        'https'
+    ]
 };
 
 export default options;