+++ src/extension.ts
@@ -10,6 +10,7 @@ import { activate as activateDiagnostic } from './diagnostic';
 import { activate as activateHover } from './hover';
 import { activate as activateNextError } from './next-error';
 import { activate as activateRefresh } from './refresh';
+import { activate as activateShare } from './share';
 import { activate as activateSnippets } from './snippet';
 import { activate as activateSubscriptions } from './subscriptions';
 import { activate as activateSymbol } from './symbol';
@@ -43,6 +44,7 @@ export function activate(context: ExtensionContext) {
   activateAPI(context);
   activateSnippets(context);
   activateColor(context);
+  activateShare(context);
 }
 
 export function deactivate() {}